Migrate runner unit tests to module
