+++ tests/unit/runners/test_ui_rgw.py
@@ -90,174 +90,3 @@ class TestRadosgw():
 
         assert result == rg.credentials
 
-    @patch('salt.client.LocalClient', autospec=True)
-    @patch('os.path.isfile', new=f_os.path.isfile)
-    @patch('os.path.exists', new=f_os.path.exists)
-    @patch('salt.config.client_config', autospec=True)
-    @patch('salt.utils.master.MasterPillarUtil', autospec=True)
-    @patch('__builtin__.open', new=f_open)
-    @patch('glob.glob', new=f_glob.glob)
-    def test_urls_dedicated_node(self, masterpillarutil, config, localclient):
-        result = {'urls': ["http://rgw1:7480/admin"],
-                  'access_key': None,
-                  'secret_key': None,
-                  'user_id': None,
-                  'success': False}
-
-        localclient().cmd.return_value = {
-            'rgw1': {
-                'fqdn': 'rgw1'
-            }
-        }
-
-        fs.CreateFile('cache/client.rgw.rgw1.json',
-            contents='''[client.rgw.rgw1]\nkey = 12345\ncaps mon = "allow rwx"\ncaps osd = "allow rwx"\n''')
-        fs.CreateFile('/srv/salt/ceph/configuration/files/ceph.conf.rgw',
-            contents='''[client.rgw.rgw1]\nkey = 12345\ncaps mon = "allow rwx"\ncaps osd = "allow rwx"\n''')
-
-        rg = ui_rgw.Radosgw(pathname="cache")
-        fs.RemoveFile('cache/client.rgw.rgw1.json')
-        fs.RemoveFile('/srv/salt/ceph/configuration/files/ceph.conf.rgw')
-
-        assert result == rg.credentials
-
-    @patch('salt.client.LocalClient', autospec=True)
-    @patch('os.path.isfile', new=f_os.path.isfile)
-    @patch('os.path.exists', new=f_os.path.exists)
-    @patch('salt.config.client_config', autospec=True)
-    @patch('salt.utils.master.MasterPillarUtil', autospec=True)
-    @patch('__builtin__.open', new=f_open)
-    @patch('glob.glob', new=f_glob.glob)
-    def test_urls_dedicated_node_with_ssl(self, masterpillarutil, config, localclient):
-        result = {'urls': ["https://rgw1:443/admin"],
-                  'access_key': None,
-                  'secret_key': None,
-                  'user_id': None,
-                  'success': False}
-
-        localclient().cmd.return_value = {
-            'rgw1': {
-                'fqdn': 'rgw1'
-            }
-        }
-
-        fs.CreateFile('cache/client.rgw.rgw1.json')
-        fs.CreateFile('/srv/salt/ceph/configuration/files/ceph.conf.rgw',
-            contents='''[client.rgw.rgw1]\nrgw_frontends = civetweb port=443s ssl_certificate=/etc/ceph/private/keyandcert.pem\nkey = 12345\ncaps mon = "allow rwx"\ncaps osd = "allow rwx"\n''')
-
-        rg = ui_rgw.Radosgw(pathname="cache")
-        fs.RemoveFile('cache/client.rgw.rgw1.json')
-        fs.RemoveFile('/srv/salt/ceph/configuration/files/ceph.conf.rgw')
-
-        assert result == rg.credentials
-
-    @patch('salt.client.LocalClient', autospec=True)
-    @patch('os.path.isfile', new=f_os.path.isfile)
-    @patch('os.path.exists', new=f_os.path.exists)
-    @patch('salt.config.client_config', autospec=True)
-    @patch('salt.utils.master.MasterPillarUtil', autospec=True)
-    @patch('__builtin__.open', new=f_open)
-    @patch('glob.glob', new=f_glob.glob)
-    def test_urls_dedicated_node_with_admin_entry(self, masterpillarutil, config, localclient):
-        result = {'urls': ["https://rgw1:443/sys"],
-                  'access_key': None,
-                  'secret_key': None,
-                  'user_id': None,
-                  'success': False}
-
-        localclient().cmd.return_value = {
-            'rgw1': {
-                'fqdn': 'rgw1'
-            }
-        }
-
-        fs.CreateFile('cache/client.rgw.rgw1.json')
-        fs.CreateFile('/srv/salt/ceph/configuration/files/ceph.conf.rgw',
-            contents='''[client.rgw.rgw1]\nrgw_frontends = civetweb port=443s\nkey = 12345\ncaps mon = "allow rwx"\ncaps osd = "allow rwx"\nrgw admin entry = sys\n''')
-
-        rg = ui_rgw.Radosgw(pathname="cache")
-        fs.RemoveFile('cache/client.rgw.rgw1.json')
-        fs.RemoveFile('/srv/salt/ceph/configuration/files/ceph.conf.rgw')
-
-        assert result == rg.credentials
-
-    @patch('salt.client.LocalClient', autospec=True)
-    @patch('os.path.isfile', new=f_os.path.isfile)
-    @patch('os.path.exists', new=f_os.path.exists)
-    @patch('salt.config.client_config', autospec=True)
-    @patch('salt.utils.master.MasterPillarUtil', autospec=True)
-    @patch('__builtin__.open', new=f_open)
-    @patch('glob.glob', new=f_glob.glob)
-    def test_urls_shared_node(self, masterpillarutil, config, localclient):
-        result = {'urls': ["http://rgw:7480/admin"],
-                  'access_key': None,
-                  'secret_key': None,
-                  'user_id': None,
-                  'success': False}
-
-        localclient().cmd.return_value = {
-            'rgw1': {
-                'fqdn': 'rgw'
-            }
-        }
-
-        fs.CreateFile('cache/client.rgw.json',
-            contents='''[client.rgw]\nkey = 12345\ncaps mon = "allow rwx"\ncaps osd = "allow rwx"\n''')
-        fs.CreateFile('/srv/salt/ceph/configuration/files/ceph.conf.rgw',
-            contents='''[client.rgw.rgw1]\nkey = 12345\ncaps mon = "allow rwx"\ncaps osd = "allow rwx"\n''')
-
-        rg = ui_rgw.Radosgw(pathname="cache")
-        fs.RemoveFile('cache/client.rgw.json')
-        fs.RemoveFile('/srv/salt/ceph/configuration/files/ceph.conf.rgw')
-
-        assert result == rg.credentials
-
-    @patch('salt.client.LocalClient', autospec=True)
-    @patch('os.path.isfile', new=f_os.path.isfile)
-    @patch('os.path.exists', new=f_os.path.exists)
-    @patch('salt.config.client_config', autospec=True)
-    @patch('salt.utils.master.MasterPillarUtil', autospec=True)
-    @patch('__builtin__.open', new=f_open)
-    @patch('glob.glob', new=f_glob.glob)
-    def test_urls_shared_node_with_ssl(self, masterpillarutil, config, localclient):
-        result = {'urls': ["https://rgw:443/admin"],
-                  'access_key': None,
-                  'secret_key': None,
-                  'user_id': None,
-                  'success': False}
-
-        localclient().cmd.return_value = {
-            'rgw1': {
-                'fqdn': 'rgw'
-            }
-        }
-
-        fs.CreateFile('cache/client.rgw.json')
-        fs.CreateFile('/srv/salt/ceph/configuration/files/ceph.conf.rgw',
-            contents='''[client.rgw.rgw1]\nrgw_frontends = civetweb port=443s ssl_certificate=/etc/ceph/private/keyandcert.pem\nkey = 12345\ncaps mon = "allow rwx"\ncaps osd = "allow rwx"\n''')
-
-        rg = ui_rgw.Radosgw(pathname="cache")
-        fs.RemoveFile('cache/client.rgw.json')
-        fs.RemoveFile('/srv/salt/ceph/configuration/files/ceph.conf.rgw')
-
-        assert result == rg.credentials
-
-
-    @patch('salt.client.LocalClient', autospec=True)
-    @patch('salt.config.client_config', autospec=True)
-    @patch('salt.utils.master.MasterPillarUtil', autospec=True)
-    @patch('__builtin__.open', new=f_open)
-    @patch('glob.glob', new=f_glob.glob)
-    def test_urls_endpoint_defined(self, masterpillarutil, config, localclient):
-        result = {'urls': ["http://abc.def/admin"],
-                  'access_key': None,
-                  'secret_key': None,
-                  'user_id': None,
-                  'success': False}
-
-        mpu = masterpillarutil.return_value
-        mpu.get_minion_pillar.return_value = { "minionA": { "rgw_endpoint": "http://abc.def/admin" }}
-        rg = ui_rgw.Radosgw(pathname="cache")
-
-        assert result == rg.credentials
-
