+++ tox.ini
@@ -20,6 +20,7 @@ deps =
     pytest
     salttesting
     configobj
+    boto
 
 [testenv:lint]
 commands = pylint --rcfile=.pylintrc --jobs=5 srv/ etc/
